fix: Resolve issue with multiple reactions per user and bump version to 0.2.6
This commit fixes an issue where a user could add multiple reactions to a single message. Now, a user can only have one reaction per message. If a user selects a new reaction, it replaces the old one. If they select the same reaction, it is removed.

The following changes were made:

- Updated the logic in `lib/src/controllers/reactions_controller.dart` to allow only one reaction per user per message.
- Bumped the package version in `pubspec.yaml` from 0.2.5 to 0.2.6.
- Updated the package version in `example/pubspec.lock` from 0.1.0 to 0.2.6.
- Updated the CHANGELOG.md to reflect the new version and changes.
- Updated the `emoji_picker_flutter` checksum in `example/ios/Podfile.lock`.